Fix blocks not sending

+++ src/com/huskehhh/fakeblock/objects/Wall.java
@@ -86,12 +86,12 @@ public class Wall {
     public String convertToString() {
         String sep = ",";
 
-        double x = loc1.getX();
-        double x1 = loc2.getX();
-        double y = loc1.getY();
-        double y1 = loc2.getY();
-        double z = loc1.getZ();
-        double z1 = loc2.getZ();
+        int x = (int) loc1.getX();
+        int x1 = (int) loc2.getX();
+        int y = (int) loc1.getY();
+        int y1 = (int) loc2.getY();
+        int z = (int) loc1.getZ();
+        int z1 = (int) loc2.getZ();
 
         String worldname = loc1.getWorld().getName();
